Add a separate module for Objective-C test fixtures
We need to define classes to write unit tests against. These must be written in Objective-C because it's not possible to model all Objective-C features in Swift (e.g. atomic properties).

The Objective-C code must reside in a separate module to appease the Swift Package Manager. Likewise, this module should be a test-only dependency, but SwiftPM doesn't support test-only dependencies at the moment (as of Swift 3.0.2).